requirements.txt
aiogram==2.11.2
aiohttp==3.7.4
async-timeout==3.0.1
asyncio==3.4.3
asyncpg==0.22.0
attrs==20.3.0
Babel==2.9.0
certifi==2020.12.5
chardet==3.0.4
idna==2.10
multidict==5.1.0
pathlib==1.0
python-dotenv==0.15.0
pytz==2021.1
requests==2.25.1
typing-extensions==3.7.4.3
urllib3==1.26.3
yarl==1.6.3
